Update: Navigator injected on TvShowCatalogPresenter
* All the navigation logic has been moved into the navigator, even the fragment related one.
